bot/pyproject.toml
[tool.poetry]
name = "telegram-mcbe"
version = "0.2.25-beta"
description = "ВК 🤝 MCPE/BE"
authors = ["Taskovich"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.111.0"
uvicorn = "0.29.0"
aiosqlite = "0.20.0"
python-dotenv = "1.0.1"
bcrypt = "4.1.3"
loguru = "0.7.2"
asyncio = "3.4.3"
pyyaml = "6.0.1"
aio-mc-rcon = "3.2.2"
aiogram = "3.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
